requirements.txt
# azure_storage==0.37.0
beautifulsoup4==4.12.2
loguru==0.7.2
openai==0.28.1
# PyQt5==5.15.10
# PyQt5_sip==12.13.0
python-dotenv==1.0.0
Requests==2.31.0
tenacity==8.2.3
toml==0.10.2
azure-cosmos==4.5.1
azure.identity==1.15.0
pandas
bs4
tiktoken
matplotlib